refactor(tailwind-config): Introduce & use base config to be consumed by frontend & serlo-editor-for-edusharing
