+++ shared/README.md
@@ -1,0 +1,1 @@
+For shared files that are also consumed by other repositories like serlo-editor-for-edusharing
